chat-with-video/requirements.txt
langchain
langchain_community
langchain-openai
openai
python-dotenv
ffmpeg
requests
streamlit
ffmpeg-python
langchain-pinecone